oapenwb-javalin: - Moved UserRevisionEntity. - Added new column platform to UserRevisionEntity. - Extended UserRevisionListener to fill new platform on revision entity. - Minor optimizations in HibernateConfiguratorImpl.
